Q-Q plot
This Q-Q plot is representative of a lot of the data processing steps that occurred while cleaning the data. First, I examined data shape using a correlation matrix, determining that correlations did exist and that a linear regression would be the best fit. I then looked for multicollinearity between data points, eliminating those that were too highly correlated. The Q-Q plot, unfortunately, showed that the data was not quite as normal as could be desired, and also indicates the presence of some extreme outliers. However, because these outliers are in large part what is being examined, I decided to keep them in the dataset. I did take the square root of the column failing normality, and this actually did fix the issue to a degree, but I did not include the processing step in the final data as it was making analysis too difficult. In an actual publication, though, it might be important to replicate as a step to show that results are not changed.