Rworldmap
This map shows countries with a high percentage of GDP being made up for by natural resource exports. The highest outlier on this map is the DRC, although the key does not show it and could be updated to improve the effectiveness of the visualization. In general, this is just a very cool package that binds country codes to a world map, and I wanted to include it somewhere in the project.