Linear Regression
This linear regression has a fairly good R-squared value for exploratory analysis of 0.4. The axes are natural resource exports as a percent of GDP (x-axis), as impacted by political rights index, health spending as a percent of GDP, military spending as a percent of GDP, globalization index, and external interventions index, the metrics with an acceptable p-value. One limitation of this regression is homoscedasticity, but this was addressed by filling in some missing countries to try to even out the distribution. It is also important to note that the clustered data around 0 on the x axis are in fact values of 0 or close to it, and not null values being graphed regardless, so they should be included. What this regression shows is that, holding globalization index and external interventions constant, military spending is increased, health spending is decreased, and political rights are decreased by having high exports of natural resources, adding support to the resource trap theory.